created a registry module for connectors
