+++ rllib/connectors/action/clip.py
@@ -1,10 +1,6 @@
 from typing import Any
 
-from ray.rllib.connectors.connector import (
-    ActionConnector,
-    ConnectorContext,
-    register_connector,
-)
+from ray.rllib.connectors.connector import ActionConnector, ConnectorContext
 from ray.rllib.utils.spaces.space_utils import clip_action, get_base_struct_from_space
 from ray.rllib.utils.typing import ActionConnectorDataType
 from ray.util.annotations import PublicAPI
@@ -36,6 +32,3 @@ class ClipActionsConnector(ActionConnector):
     @staticmethod
     def from_state(ctx: ConnectorContext, params: Any):
         return ClipActionsConnector(ctx)
-
-
-register_connector(ClipActionsConnector.__name__, ClipActionsConnector)
