+++ rllib/connectors/action/normalize.py
@@ -1,10 +1,6 @@
 from typing import Any
 
-from ray.rllib.connectors.connector import (
-    ActionConnector,
-    ConnectorContext,
-    register_connector,
-)
+from ray.rllib.connectors.connector import ActionConnector, ConnectorContext
 from ray.rllib.utils.spaces.space_utils import (
     get_base_struct_from_space,
     unsquash_action,
@@ -39,6 +35,3 @@ class NormalizeActionsConnector(ActionConnector):
     @staticmethod
     def from_state(ctx: ConnectorContext, params: Any):
         return NormalizeActionsConnector(ctx)
-
-
-register_connector(NormalizeActionsConnector.__name__, NormalizeActionsConnector)
